[mapleall]Internal commit msg:     [mapleall]Static check alarm cleaning(4)

+++ src/mapleall/maple_be/include/cg/aarch64/aarch64_abi.h
@@ -96,7 +96,7 @@ class ParmLocator {
 
  private:
   BECommon &beCommon;
-  int32 paramNum           = 0;  /* number of all types of parameters processed so far */
+  uint64 paramNum           = 0;  /* number of all types of parameters processed so far */
   int32 nextGeneralRegNO   = 0;  /* number of integer parameters processed so far */
   uint32 nextFloatRegNO    = 0;  /* number of float parameters processed so far */
   int32 nextStackArgAdress = 0;
